move dummy data to separate file
