+++ lib/dai-plugin-mcd/test/CdpManager.spec.js
@@ -1,7 +1,7 @@
 import { mcdMaker, setDebtCeiling, setupCollateral } from './helpers';
 import { ServiceRoles } from '../src/constants';
 import { ETH, MDAI } from '../src';
-import { dummyEventData, formattedDummyEventData } from './ManagedCdp.spec.js';
+import { dummyEventData, formattedDummyEventData } from './fixtures';
 
 // FIXME we won't be able to reach into @makerdao/dai internals like this when
 // this plugin is moved into its own module...
